Initial stages of Redis Monitor plugin setup
Info requests working, all other commented out for now. Alsoe added docopt for help running the program. Also added some plugins for handling outbound Kafka messages and the info plugin itself.
