+++ redis-monitor/settings.py
@@ -4,3 +4,5 @@ REDIS_PORT = 6379
 
 KAFKA_HOSTS = ''
 KAFKA_TOPIC_PREFIX = 'demo'
+
+PLUGINS = {}
